refactor: extract shared API route helpers to reduce duplication
- Create /src/lib/api/route-helpers.ts with common auth/access patterns
- Refactor profile route to use shared withAuthAndAccess helper
- Remove duplicate authentication and validation code
- Eliminate the helpers.ts file to reduce code duplication
- All tests passing, build successful

Addresses Codacy duplication issues for PR #211
